Add files via upload
Parsed using nlohmann parser; added print function to the course class for easier use. The JSON cs_major file has more attributes than what was in the Course class; i.e missing: department, number. Didn't include it to the Course class, but this can be extended to add all of the attributes into the class if need be.